+++ src/components/FooterAppBar/FooterAppBar.test.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { render, screen } from '@testing-library/react';
+import { FooterAppBar } from './FooterAppBar';
+
+describe('Pil component', () => {
+  it('it has the correct text ', () => {
+    render(<FooterAppBar />);
+    expect(false).toBeTruthy();
+  });
+});
