+++ src/components/FooterAppBar/index.ts
@@ -1,0 +1,1 @@
+export { FooterAppBar } from './FooterAppBar';
